Refactored MaskRCNN to RetinaMask (#360)
