+++ deepcell/model_zoo/__init__.py
@@ -40,8 +40,8 @@ from deepcell.model_zoo.retinanet import RetinaNet
 from deepcell.model_zoo.retinanet import retinanet
 from deepcell.model_zoo.retinanet import retinanet_bbox
 
-from deepcell.model_zoo.maskrcnn import RetinaMask
-from deepcell.model_zoo.maskrcnn import retinanet_mask
+from deepcell.model_zoo.retinamask import RetinaMask
+from deepcell.model_zoo.retinamask import retinanet_mask
 
 from deepcell.model_zoo.fpn import FPNet
